src/test/java/com/github/jcustenborder/kafka/connect/solr/SolrInputDocumentBuilderTest.java
/**
 * Copyright © 2016 Jeremy Custenborder (jcustenborder@gmail.com)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package com.github.jcustenborder.kafka.connect.solr;

import org.apache.kafka.connect.data.Field;
import org.apache.solr.common.SolrInputDocument;
import org.junit.jupiter.api.Test;

import java.util.List;
import java.util.Map;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;

public class SolrInputDocumentBuilderTest {

  @Test
  public void buildMap() {
    Records.MapTestCase mapTestCase = Records.map();

    SolrInputDocument document = SolrInputDocumentBuilder.build(mapTestCase.record);
    assertNotNull(document, "document should not be null.");
    for (Map.Entry<String, Object> kvp : mapTestCase.map.entrySet()) {
      assertEquals(kvp.getValue(), document.getFieldValue(kvp.getKey()), kvp.getKey() + " does not match.");
    }
  }


  @Test
  public void buildStruct() {
    Records.StructTestCase structTestCase = Records.struct();
    SolrInputDocument document = SolrInputDocumentBuilder.build(structTestCase.record);
    assertNotNull(document, "document should not be null.");
    List<Field> fields = structTestCase.struct.schema().fields();
    for (Field field : fields) {
      Object value = structTestCase.struct.get(field);
      assertEquals(value, document.getFieldValue(field.name()), field.name() + " does not match.");
    }
  }

}
